Moved the TurretType enum in the Turret class to above the Turret class, allowing it to be used by the Game1 class.
Added an X and Y parameter to the Turret constructor.

Attempted to test the turret.Draw() method with a test image. Currently unsuccesful.

I theorize that the reason the image isn't appearing is not because of a coding error explicitly, but rather that the image I chose was too big for the rectangle in the Turret class. I will try to remedy this at either our next meeting or on my own time.